code restructuring as per typical golang projects
