+++ docs/how-it-works.md
