+++ docs/img/kube-router-arch.png
